% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pits_power_result}
\alias{print.pits_power_result}
\title{Print method for PITS power results}
\usage{
\method{print}{pits_power_result}(x, ...)
}
\arguments{
\item{x}{A \code{pits_power_result} object.}

\item{...}{Ignored.}
}
\value{
Invisibly \code{x}.
}
\description{
Displays a formatted summary of the output from \code{\link[=calculate_power]{calculate_power()}} or
\code{\link[=calculate_power_multi]{calculate_power_multi()}}.
}
